refactor opencode
