+++ src/providers/index.ts
@@ -4,7 +4,7 @@
 
 export { BaseEmbeddedProvider } from "./embedded/base.js";
 export { ClaudeCodeProvider } from "./embedded/claudeCode.js";
-export { OpenCodeProvider } from "./embedded/openCode.js";
+export { OpenCodeProvider } from "./embedded/opencode.js";
 export {
 	agentProviderFactory,
 	DefaultAgentProviderFactory,
